#33 wip move search to currency lib
